Moved all loading of Parse data to the LoadActivity.
When you click LoadActivity now if you are logged into Facebook, it will query Parse and download all songs/tracks for that user.

It will match ONLY SONGS currently against Songs in the SQLite DB.  If the song already exists, it will ignore the tracks.  If there was no song in internal DB it will create the directory and store all of the tracks from Parse to internal memory.

If Guest, it will only check internal memory.